올림

+++ Assets/02.Scripts/08.Excel/EquipmentLvTierTemplate.cs
@@ -5,7 +5,7 @@ public class EquipmentLvTierTemplate
 {
     public int Index;
 
-    public float Common, Magic, Rare, Heroic, Legendary, Unique, Mythic, Ancient, Abyssal, Genesis;
+    public int Common, Magic, Rare, Heroic, Legendary, Unique, Mythic, Ancient, Abyssal, Genesis;
 
     public EquipmentLvTierTemplate() { }
 
@@ -20,15 +20,15 @@ public class EquipmentLvTierTemplate
 
         Index = Convert.ToInt32(listValue[wCount++]);
 
-        Common = Convert.ToSingle(listValue[wCount++]);
-        Magic = Convert.ToSingle(listValue[wCount++]);
-        Rare = Convert.ToSingle(listValue[wCount++]);
-        Heroic = Convert.ToSingle(listValue[wCount++]);
-        Legendary = Convert.ToSingle(listValue[wCount++]);
-        Unique = Convert.ToSingle(listValue[wCount++]);
-        Mythic = Convert.ToSingle(listValue[wCount++]);
-        Ancient = Convert.ToSingle(listValue[wCount++]);
-        Abyssal = Convert.ToSingle(listValue[wCount++]);
-        Genesis = Convert.ToSingle(listValue[wCount++]);
+        Common = Convert.ToInt32(listValue[wCount++]);
+        Magic = Convert.ToInt32(listValue[wCount++]);
+        Rare = Convert.ToInt32(listValue[wCount++]);
+        Heroic = Convert.ToInt32(listValue[wCount++]);
+        Legendary = Convert.ToInt32(listValue[wCount++]);
+        Unique = Convert.ToInt32(listValue[wCount++]);
+        Mythic = Convert.ToInt32(listValue[wCount++]);
+        Ancient = Convert.ToInt32(listValue[wCount++]);
+        Abyssal = Convert.ToInt32(listValue[wCount++]);
+        Genesis = Convert.ToInt32(listValue[wCount++]);
     }
 }
